refactor: 파일 크기 변환 로직 formatFileSize 유틸 함수로 추출 및 재사용 적용(#356)
